# byzantine.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "byzantine.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/broker/echo": {
      "post": {
        "operationId": "Echo",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/byzantineEchoResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/byzantinePublication"
            }
          }
        ],
        "tags": [
          "Broker"
        ]
      }
    },
    "/broker/pull/{PublisherID}": {
      "post": {
        "operationId": "Receive",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/byzantinePubResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "PublisherID",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/byzantinePublication"
            }
          }
        ],
        "tags": [
          "Broker"
        ]
      }
    },
    "/broker/push": {
      "get": {
        "operationId": "Push",
        "responses": {
          "200": {
            "description": "(streaming responses)",
            "schema": {
              "$ref": "#/definitions/byzantinePublication"
            }
          }
        },
        "parameters": [
          {
            "name": "PublisherID",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "BrokerID",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Broker"
        ]
      }
    },
    "/broker/ready": {
      "post": {
        "operationId": "Ready",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/byzantineReadyResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/byzantinePublication"
            }
          }
        ],
        "tags": [
          "Broker"
        ]
      }
    },
    "/broker/register/{PoolID}": {
      "post": {
        "operationId": "RegisterSubscriber",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/byzantineReadyResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "PoolID",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/byzantineSubscriber"
            }
          }
        ],
        "tags": [
          "Broker"
        ]
      }
    },
    "/broker/subscribers": {
      "get": {
        "operationId": "GetSubscribers",
        "responses": {
          "200": {
            "description": "(streaming responses)",
            "schema": {
              "$ref": "#/definitions/byzantineSubscriber"
            }
          }
        },
        "parameters": [
          {
            "name": "Address",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "PoolID",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "DestinationDistinguishment",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Broker"
        ]
      }
    },
    "/subscriber/subscribe": {
      "post": {
        "operationId": "Subscribe",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/byzantineReadyResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/byzantineSubRequest"
            }
          }
        ],
        "tags": [
          "Subscribe"
        ]
      }
    }
  },
  "definitions": {
    "byzantineChainMAC": {
      "type": "object",
      "properties": {
        "From": {
          "type": "string"
        },
        "To": {
          "type": "string"
        },
        "MAC": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "byzantineChainResponse": {
      "type": "object",
      "properties": {
        "Valid": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "byzantineEchoResponse": {
      "type": "object",
      "properties": {
        "Hello": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "byzantinePubResponse": {
      "type": "object",
      "properties": {
        "Success": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "byzantinePublication": {
      "type": "object",
      "properties": {
        "PubType": {
          "type": "integer",
          "format": "int64"
        },
        "PublisherID": {
          "type": "string",
          "format": "uint64"
        },
        "PublicationID": {
          "type": "string",
          "format": "int64"
        },
        "TopicID": {
          "type": "string",
          "format": "uint64"
        },
        "BrokerID": {
          "type": "string",
          "format": "uint64"
        },
        "Contents": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "byte"
          }
        },
        "MAC": {
          "type": "string",
          "format": "byte"
        },
        "ChainMACs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/byzantineChainMAC"
          }
        }
      },
      "description": "/ Publication is the message sent out to the brokers, who pass it along to the subscribers."
    },
    "byzantineReadyResponse": {
      "type": "object",
      "properties": {
        "Ready": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "byzantineSubRequest": {
      "type": "object",
      "properties": {
        "PublisherID": {
          "type": "string",
          "format": "uint64"
        },
        "BrokerID": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "byzantineSubscriber": {
      "type": "object",
      "properties": {
        "Address": {
          "type": "string"
        },
        "PoolID": {
          "type": "string",
          "format": "uint64"
        },
        "DestinationDistinguishment": {
          "type": "string",
          "format": "uint64"
        }
      },
      "description": "/ Subscriber defines a subscriber within a subscriber pool."
    }
  }
}